# Conflicts:
#   teamup-auth/src/main/java/ru/team/up/auth/controller/authController/AuthController.java
#   teamup-input/src/main/java/ru/team/up/input/controller/privateController/AdminController.java
#   teamup-input/src/main/java/ru/team/up/input/controller/privateController/EmailUserMessageNotificatorController.java
#   teamup-input/src/main/java/ru/team/up/input/controller/privateController/ModeratorController.java

+++ teamup-input/src/main/java/ru/team/up/input/controller/privateController/AdminController.java
@@ -16,6 +16,7 @@ import ru.team.up.core.monitoring.service.MonitorProducerService;
 import ru.team.up.core.service.AdminService;
 import ru.team.up.dto.ControlDto;
 import ru.team.up.dto.UserDto;
+import ru.team.up.sup.service.ParameterService;
 
 import javax.validation.constraints.NotNull;
 import java.util.HashMap;
@@ -45,6 +46,10 @@ public class AdminController {
     @GetMapping
     public ResponseEntity<List<Account>> getAllAdmins() {
         log.debug("Старт метода ResponseEntity<List<Admin>> getAllAdmins()");
+        if (!ParameterService.getAllAdminsEnabled.getValue()) {
+            log.debug("Метод getAllAdmins выключен параметром getAllAdminsEnabled = false");
+            throw new RuntimeException("Method getAllAdmins is disabled by parameter getAllAdminsEnabled");
+        }
 
         List<Account> allAdmins = adminService.getAllAdmins();
 
@@ -75,6 +80,13 @@ public class AdminController {
         log.debug("Старт метода ResponseEntity<Admin> getOneAdmin(@PathVariable Long id) с параметром {}", id);
         Account admin = adminService.getOneAdmin(id);
         ResponseEntity<Account> responseEntity = ResponseEntity.ok(admin);
+
+        if (!ParameterService.getOneAdminEnabled.getValue()) {
+            log.debug("Метод getOneAdmin выключен параметром getOneAdminEnabled = false");
+            throw new RuntimeException("Method getOneAdmin is disabled by parameter getOneAdminEnabled");
+        }
+
+        ResponseEntity<Account> responseEntity = ResponseEntity.ok(admin);
         log.debug("Получили ответ {}", responseEntity);
 
         Map<String, Object> monitoringParameters = new HashMap<>();
@@ -100,6 +112,10 @@ public class AdminController {
     @PostMapping
     public ResponseEntity<Account> createAdmin(@RequestBody @NotNull Admin adminCreate) {
         log.debug("Старт метода ResponseEntity<Admin> createAdmin(@RequestBody @NotNull Admin admin) с параметром {}", adminCreate);
+        if (!ParameterService.createAdminEnabled.getValue()) {
+            log.debug("Метод createAdmin выключен параметром createAdminEnabled = false");
+            throw new RuntimeException("Method createAdmin is disabled by parameter createAdminEnabled");
+        }
         ResponseEntity<Account> responseEntity =
                 new ResponseEntity<>(adminService.saveAdmin(adminCreate), HttpStatus.CREATED);
         log.debug("Получили ответ {}", responseEntity);
@@ -127,6 +143,12 @@ public class AdminController {
     @PutMapping("/{id}")
     public ResponseEntity<Account> updateAdmin(@PathVariable Long id, @RequestBody @NotNull Admin admin) {
         log.debug("Старт метода ResponseEntity<Admin> updateAdmin(@RequestBody @NotNull Admin admin) с параметром {}", admin);
+
+        if (!ParameterService.updateAdminEnabled.getValue()) {
+            log.debug("Метод updateAdmin выключен параметром updateAdminEnabled = false");
+            throw new RuntimeException("Method updateAdmin is disabled by parameter updateAdminEnabled");
+        }
+
         log.debug("Проверка наличия обновляемого администратора в БД");
         if (adminService.existsById(id) && id.equals(admin.getId())) {
             ResponseEntity<Account> responseEntity = ResponseEntity.ok(adminService.updateAdmin(admin));
@@ -158,6 +180,12 @@ public class AdminController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Admin> deleteAdmin(@PathVariable Long id) {
         log.debug("Старт метода ResponseEntity<Admin> updateAdmin(@RequestBody @NotNull Admin admin) с параметром {}", id);
+
+        if (!ParameterService.deleteAdminFromAdminControllerEnabled.getValue()) {
+            log.debug("Метод deleteAdmin выключен параметром deleteAdminFromAdminControllerEnabled = false");
+            throw new RuntimeException("Method deleteAdmin is disabled by parameter deleteAdminFromAdminControllerEnabled");
+        }
+
         adminService.deleteAdmin(id);
 
         ResponseEntity<Admin> responseEntity = new ResponseEntity<>(HttpStatus.OK);
